+++ src/client/java/io/github/smajloslovakian/smoothscroll/CreativeModeInventoryScreenDuck.java
@@ -5,4 +5,5 @@ import net.minecraft.client.gui.GuiGraphics;
 public interface CreativeModeInventoryScreenDuck {
     void enMask(GuiGraphics graphics);
     void deMask(GuiGraphics graphics);
+    boolean isMouseInbounds();
 }
