+++ src/client/java/io/github/smajloslovakian/smoothscroll/menu/ModMenuImpl.java
@@ -6,6 +6,6 @@ import com.terraformersmc.modmenu.api.ModMenuApi;
 public class ModMenuImpl implements ModMenuApi {
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return parent -> {return new Inspiration(parent);};
+        return parent -> {return new SmoothScrollConfigScreen(parent);};
     }
 }
