+++ src/main/resources/assets/smoothscroll/lang/en_us.json
@@ -20,6 +20,7 @@
   "smoothscroll.config.chat.suggestion_speed": "Suggestion Speed",
   "smoothscroll.config.creative_screen": "Creative Screen",
   "smoothscroll.config.creative_screen.smoothness": "Smoothness",
+  "smoothscroll.config.creative_screen.speed": "Speed",
   "smoothscroll.config.entry_list": "Entry List",
   "smoothscroll.config.entry_list.smoothness": "Smoothness",
   "smoothscroll.config.entry_list.speed": "Speed",
